Change the story file name to include stories
When we move to version 7+, this will be updated to mgt-component.js from mgt-component.stories.js
