+++ stories/components/people/people.stories.js
