+++ stories/components/peoplePicker/peoplePicker.stories.js
