+++ stories/components/preview/searchBox/searchBox.stories.js
